pipeline/ingestion/requirements.txt
requests==2.31.0
psycopg[binary]>=3.1.0
python-dotenv==1.0.0